Refactor abstraction alerts folder structure
https://eaflood.atlassian.net/browse/WATER-5630

This change moves abstraction alerts services and presenters into the abstraction alerts folder.
