+++ app/controllers/notices-setup.controller.js
@@ -9,13 +9,13 @@ const InitiateSessionService = require('../services/notices/setup/initiate-sessi
 const ProcessAddRecipientService = require('../services/notices/setup/process-add-recipient.service.js')
 const ProcessDownloadRecipientsService = require('../services/notices/setup/process-download-recipients.service.js')
 const ProcessPreviewPaperReturnService = require('../services/notices/setup/process-preview-paper-return.service.js')
-const ProcessRemoveThresholdService = require('../services/notices/setup/process-remove-threshold.service.js')
-const SubmitAlertEmailAddressService = require('../services/notices/setup/submit-alert-email-address.service.js')
-const SubmitAlertThresholdsService = require('../services/notices/setup/submit-alert-thresholds.service.js')
-const SubmitAlertTypeService = require('../services/notices/setup/submit-alert-type.service.js')
-const SubmitCancelAlertsService = require('../services/notices/setup/submit-cancel-alerts.service.js')
+const ProcessRemoveThresholdService = require('../services/notices/setup/abstraction-alerts/process-remove-threshold.service.js')
+const SubmitAlertEmailAddressService = require('../services/notices/setup/abstraction-alerts/submit-alert-email-address.service.js')
+const SubmitAlertThresholdsService = require('../services/notices/setup/abstraction-alerts/submit-alert-thresholds.service.js')
+const SubmitAlertTypeService = require('../services/notices/setup/abstraction-alerts/submit-alert-type.service.js')
+const SubmitCancelAlertsService = require('../services/notices/setup/abstraction-alerts/submit-cancel-alerts.service.js')
 const SubmitCancelService = require('../services/notices/setup/submit-cancel.service.js')
-const SubmitCheckLicenceMatchesService = require('../services/notices/setup/submit-check-licence-matches.service.js')
+const SubmitCheckLicenceMatchesService = require('../services/notices/setup/abstraction-alerts/submit-check-licence-matches.service.js')
 const SubmitCheckNoticeTypeService = require('../services/notices/setup/submit-check-notice-type.service.js')
 const SubmitCheckService = require('../services/notices/setup/submit-check.service.js')
 const SubmitContactTypeService = require('../services/notices/setup/submit-contact-type.service.js')
@@ -26,13 +26,13 @@ const SubmitRecipientNameService = require('../services/notices/setup/submit-rec
 const SubmitRemoveLicencesService = require('../services/notices/setup/submit-remove-licences.service.js')
 const SubmitReturnsPeriodService = require('../services/notices/setup/submit-returns-period.service.js')
 const SubmitSelectRecipientsService = require('../services/notices/setup/submit-select-recipients.service.js')
-const ViewAlertEmailAddressService = require('../services/notices/setup/view-alert-email-address.service.js')
-const ViewAlertThresholdsService = require('../services/notices/setup/view-alert-thresholds.service.js')
-const ViewAlertTypeService = require('../services/notices/setup/view-alert-type.service.js')
+const ViewAlertEmailAddressService = require('../services/notices/setup/abstraction-alerts/view-alert-email-address.service.js')
+const ViewAlertThresholdsService = require('../services/notices/setup/abstraction-alerts/view-alert-thresholds.service.js')
+const ViewAlertTypeService = require('../services/notices/setup/abstraction-alerts/view-alert-type.service.js')
 const ViewCancelService = require('../services/notices/setup/view-cancel.service.js')
-const ViewCancelAlertsService = require('../services/notices/setup/view-cancel-alerts.service.js')
+const ViewCancelAlertsService = require('../services/notices/setup/abstraction-alerts/view-cancel-alerts.service.js')
 const ViewCheckService = require('../services/notices/setup/view-check.service.js')
-const ViewCheckLicenceMatchesService = require('../services/notices/setup/view-check-licence-matches.service.js')
+const ViewCheckLicenceMatchesService = require('../services/notices/setup/abstraction-alerts/view-check-licence-matches.service.js')
 const ViewCheckNoticeTypeService = require('../services/notices/setup/view-check-notice-type.service.js')
 const ViewConfirmationService = require('../services/notices/setup/view-confirmation.service.js')
 const ViewContactTypeService = require('../services/notices/setup/view-contact-type.service.js')
