+++ app/services/notices/setup/abstraction-alerts/process-remove-threshold.service.js
@@ -6,9 +6,9 @@
  * @module ProcessRemoveThresholdService
  */
 
-const FetchSessionDal = require('../../../dal/fetch-session.dal.js')
-const GeneralLib = require('../../../lib/general.lib.js')
-const { formatRestrictionType, formatValueUnit } = require('../../../presenters/base.presenter.js')
+const FetchSessionDal = require('../../../../dal/fetch-session.dal.js')
+const GeneralLib = require('../../../../lib/general.lib.js')
+const { formatRestrictionType, formatValueUnit } = require('../../../../presenters/base.presenter.js')
 
 /**
  * Orchestrates removing the licence monitoring station from the thresholds list for - `/notices/setup/{sessionId}/abstraction-alerts/remove-threshold/{licenceMonitoringStationId}` page
