+++ app/services/notices/setup/abstraction-alerts/view-alert-email-address.service.js
@@ -6,8 +6,8 @@
  * @module ViewAlertEmailAddressService
  */
 
-const AlertEmailAddressPresenter = require('../../../presenters/notices/setup/alert-email-address.presenter.js')
-const FetchSessionDal = require('../../../dal/fetch-session.dal.js')
+const AlertEmailAddressPresenter = require('../../../../presenters/notices/setup/abstraction-alerts/alert-email-address.presenter.js')
+const FetchSessionDal = require('../../../../dal/fetch-session.dal.js')
 
 /**
  * Orchestrates fetching and presenting the data for the `/notices/setup/{sessionId}/abstraction-alerts/alert-email-address` page
