+++ app/services/notices/setup/abstraction-alerts/view-cancel-alerts.service.js
@@ -6,8 +6,8 @@
  * @module ViewCancelAlertsService
  */
 
-const CancelAlertsPresenter = require('../../../presenters/notices/setup/cancel-alerts.presenter.js')
-const FetchSessionDal = require('../../../dal/fetch-session.dal.js')
+const CancelAlertsPresenter = require('../../../../presenters/notices/setup/abstraction-alerts/cancel-alerts.presenter.js')
+const FetchSessionDal = require('../../../../dal/fetch-session.dal.js')
 
 /**
  * Orchestrates fetching and presenting the data for the `/notices/setup/{sessionId}/abstraction-alerts/cancel` page
