+++ app/services/notices/setup/abstraction-alerts/view-check-licence-matches.service.js
@@ -6,9 +6,9 @@
  * @module ViewCheckLicenceMatchesService
  */
 
-const CheckLicenceMatchesPresenter = require('../../../presenters/notices/setup/check-licence-matches.presenter.js')
-const FetchSessionDal = require('../../../dal/fetch-session.dal.js')
-const { readFlashNotification } = require('../../../lib/general.lib.js')
+const CheckLicenceMatchesPresenter = require('../../../../presenters/notices/setup/abstraction-alerts/check-licence-matches.presenter.js')
+const FetchSessionDal = require('../../../../dal/fetch-session.dal.js')
+const { readFlashNotification } = require('../../../../lib/general.lib.js')
 
 /**
  * Orchestrates presenting the data for the `/notices/setup/{sessionId}/abstraction-alerts/check-licence-matches` page
