+++ test/controllers/notices-setup.controller.test.js
@@ -17,13 +17,13 @@ const InitiateSessionService = require('../../app/services/notices/setup/initiat
 const ProcessAddRecipientService = require('../../app/services/notices/setup/process-add-recipient.service.js')
 const ProcessDownloadRecipientsService = require('../../app/services/notices/setup/process-download-recipients.service.js')
 const ProcessPreviewPaperReturnService = require('../../app/services/notices/setup/process-preview-paper-return.service.js')
-const ProcessRemoveThresholdService = require('../../app/services/notices/setup/process-remove-threshold.service.js')
-const SubmitAlertEmailAddressService = require('../../app/services/notices/setup/submit-alert-email-address.service.js')
-const SubmitAlertThresholdsService = require('../../app/services/notices/setup/submit-alert-thresholds.service.js')
-const SubmitAlertTypeService = require('../../app/services/notices/setup/submit-alert-type.service.js')
-const SubmitCancelAlertsService = require('../../app/services/notices/setup/submit-cancel-alerts.service.js')
+const ProcessRemoveThresholdService = require('../../app/services/notices/setup/abstraction-alerts/process-remove-threshold.service.js')
+const SubmitAlertEmailAddressService = require('../../app/services/notices/setup/abstraction-alerts/submit-alert-email-address.service.js')
+const SubmitAlertThresholdsService = require('../../app/services/notices/setup/abstraction-alerts/submit-alert-thresholds.service.js')
+const SubmitAlertTypeService = require('../../app/services/notices/setup/abstraction-alerts/submit-alert-type.service.js')
+const SubmitCancelAlertsService = require('../../app/services/notices/setup/abstraction-alerts/submit-cancel-alerts.service.js')
 const SubmitCancelService = require('../../app/services/notices/setup/submit-cancel.service.js')
-const SubmitCheckLicenceMatchesService = require('../../app/services/notices/setup/submit-check-licence-matches.service.js')
+const SubmitCheckLicenceMatchesService = require('../../app/services/notices/setup/abstraction-alerts/submit-check-licence-matches.service.js')
 const SubmitCheckNoticeTypeService = require('../../app/services/notices/setup/submit-check-notice-type.service.js')
 const SubmitCheckService = require('../../app/services/notices/setup/submit-check.service.js')
 const SubmitContactTypeService = require('../../app/services/notices/setup/submit-contact-type.service.js')
@@ -34,13 +34,13 @@ const SubmitRecipientNameService = require('../../app/services/notices/setup/sub
 const SubmitRemoveLicencesService = require('../../app/services/notices/setup/submit-remove-licences.service.js')
 const SubmitReturnsPeriodService = require('../../app/services/notices/setup/submit-returns-period.service.js')
 const SubmitSelectRecipientsService = require('../../app/services/notices/setup/submit-select-recipients.service.js')
-const ViewAlertEmailAddressService = require('../../app/services/notices/setup/view-alert-email-address.service.js')
-const ViewAlertThresholdsService = require('../../app/services/notices/setup/view-alert-thresholds.service.js')
-const ViewAlertTypeService = require('../../app/services/notices/setup/view-alert-type.service.js')
+const ViewAlertEmailAddressService = require('../../app/services/notices/setup/abstraction-alerts/view-alert-email-address.service.js')
+const ViewAlertThresholdsService = require('../../app/services/notices/setup/abstraction-alerts/view-alert-thresholds.service.js')
+const ViewAlertTypeService = require('../../app/services/notices/setup/abstraction-alerts/view-alert-type.service.js')
 const ViewCancelService = require('../../app/services/notices/setup/view-cancel.service.js')
-const ViewCancelAlertsService = require('../../app/services/notices/setup/view-cancel-alerts.service.js')
+const ViewCancelAlertsService = require('../../app/services/notices/setup/abstraction-alerts/view-cancel-alerts.service.js')
 const ViewCheckService = require('../../app/services/notices/setup/view-check.service.js')
-const ViewCheckLicenceMatchesService = require('../../app/services/notices/setup/view-check-licence-matches.service.js')
+const ViewCheckLicenceMatchesService = require('../../app/services/notices/setup/abstraction-alerts/view-check-licence-matches.service.js')
 const ViewCheckNoticeTypeService = require('../../app/services/notices/setup/view-check-notice-type.service.js')
 const ViewConfirmationService = require('../../app/services/notices/setup/view-confirmation.service.js')
 const ViewContactTypeService = require('../../app/services/notices/setup/view-contact-type.service.js')
