+++ test/services/notices/setup/abstraction-alerts/submit-cancel-alerts.service.test.js
@@ -9,18 +9,18 @@ const { describe, it, beforeEach, afterEach } = (exports.lab = Lab.script())
 const { expect } = Code
 
 // Test helpers
-const SessionModelStub = require('../../../support/stubs/session.stub.js')
+const SessionModelStub = require('../../../../support/stubs/session.stub.js')
 
-const { generateUUID } = require('../../../../app/lib/general.lib.js')
+const { generateUUID } = require('../../../../../app/lib/general.lib.js')
 
 // Things we need to stub
-const DeleteSessionDal = require('../../../../app/dal/delete-session.dal.js')
-const FetchSessionDal = require('../../../../app/dal/fetch-session.dal.js')
+const DeleteSessionDal = require('../../../../../app/dal/delete-session.dal.js')
+const FetchSessionDal = require('../../../../../app/dal/fetch-session.dal.js')
 
 // Thing under test
-const SubmitCancelAlertsService = require('../../../../app/services/notices/setup/submit-cancel-alerts.service.js')
+const SubmitCancelAlertsService = require('../../../../../app/services/notices/setup/abstraction-alerts/submit-cancel-alerts.service.js')
 
-describe('Notices - Setup - Submit Cancel Alerts service', () => {
+describe('Notices - Setup - Abstraction Alerts - Submit Cancel Alerts service', () => {
   let session
   let sessionData
