+++ test/services/notices/setup/abstraction-alerts/view-alert-type.service.test.js
@@ -9,16 +9,16 @@ const { describe, it, beforeEach, afterEach } = (exports.lab = Lab.script())
 const { expect } = Code
 
 // Test helpers
-const AbstractionAlertSessionData = require('../../../support/fixtures/abstraction-alert-session-data.fixture.js')
-const SessionModelStub = require('../../../support/stubs/session.stub.js')
+const AbstractionAlertSessionData = require('../../../../support/fixtures/abstraction-alert-session-data.fixture.js')
+const SessionModelStub = require('../../../../support/stubs/session.stub.js')
 
 // Things we need to stub
-const FetchSessionDal = require('../../../../app/dal/fetch-session.dal.js')
+const FetchSessionDal = require('../../../../../app/dal/fetch-session.dal.js')
 
 // Thing under test
-const ViewAlertTypeService = require('../../../../app/services/notices/setup/view-alert-type.service.js')
+const ViewAlertTypeService = require('../../../../../app/services/notices/setup/abstraction-alerts/view-alert-type.service.js')
 
-describe('Notices Setup - Setup - View Alert Type service', () => {
+describe('Notices Setup - Setup - Abstraction Alerts - View Alert Type service', () => {
   let session
   let sessionData
