+++ test/services/notices/setup/abstraction-alerts/view-check-licence-matches.service.test.js
@@ -9,17 +9,17 @@ const { describe, it, beforeEach, afterEach } = (exports.lab = Lab.script())
 const { expect } = Code
 
 // Test helpers
-const AbstractionAlertSessionData = require('../../../support/fixtures/abstraction-alert-session-data.fixture.js')
-const SessionModelStub = require('../../../support/stubs/session.stub.js')
-const YarStub = require('../../../support/stubs/yar.stub.js')
+const AbstractionAlertSessionData = require('../../../../support/fixtures/abstraction-alert-session-data.fixture.js')
+const SessionModelStub = require('../../../../support/stubs/session.stub.js')
+const YarStub = require('../../../../support/stubs/yar.stub.js')
 
 // Things we need to stub
-const FetchSessionDal = require('../../../../app/dal/fetch-session.dal.js')
+const FetchSessionDal = require('../../../../../app/dal/fetch-session.dal.js')
 
 // Thing under test
-const ViewCheckLicenceMatchesService = require('../../../../app/services/notices/setup/view-check-licence-matches.service.js')
+const ViewCheckLicenceMatchesService = require('../../../../../app/services/notices/setup/abstraction-alerts/view-check-licence-matches.service.js')
 
-describe('Notices - Setup - View Check Licence Matches service', () => {
+describe('Notices - Setup - Abstraction Alerts - View Check Licence Matches service', () => {
   let licenceMonitoringStations
   let session
   let sessionData
